Web/requirements.txt
flask
flask_login
flask-sqlalchemy
paho-mqtt